refactor: address review — extract scripts_parse, fix docstring, harden test isolation
- Move parse_check_errors into tool-independent mcp_server/scripts_parse.py so
  the debug_workflow service no longer imports the FastMCP tool layer (Copilot).
- Drop the stale "Signal connections" check from the debug_workflow docstring —
  it was never collected (Copilot).
- Clear the global _preflight_cache in fixture teardown too, for cross-module
  test isolation (Copilot).
