+++ mcp_server/tools/scripts.py
@@ -10,8 +10,6 @@ errors. All in the gated ``scripts`` toolset.
 
 from __future__ import annotations
 
-import re
-
 from fastmcp import FastMCP
 
 from mcp_server.bridge import Bridge
@@ -20,7 +18,6 @@ from mcp_server.config import ServerConfig
 from mcp_server.models.scripts import (
     NodeScript,
     ParseCheckResult,
-    ParseError,
     PatchScriptResult,
     ScriptContent,
     ScriptList,
@@ -28,37 +25,10 @@ from mcp_server.models.scripts import (
 )
 from mcp_server.runtime import Runner, resolve_project_dir
 from mcp_server.safety import MUTATING, READ_ONLY, PreconditionError, enforce_preconditions
+from mcp_server.scripts_parse import parse_check_errors
 from mcp_server.tools._route import route, run_or_preview
 
 SCRIPTS = {SCRIPTS_TAG}
-
-# "(res://path.gd:42)" as printed on a parse error's "at:" line.
-_CHECK_SOURCE_RE = re.compile(r"\((res://[^):]+):(\d+)\)")
-
-
-def parse_check_errors(text: str) -> list[ParseError]:
-    """Extract structured parse errors from ``--check-only`` output.
-
-    Godot prints ``SCRIPT ERROR: Parse Error: <message>`` followed by an
-    ``at: … (res://file.gd:LINE)`` line; pair them up.
-    """
-    lines = text.splitlines()
-    errors: list[ParseError] = []
-    for i, line in enumerate(lines):
-        # Only genuine parse errors — not other SCRIPT ERROR shapes (e.g. load
-        # failures) that --check-only may also print.
-        if "Parse Error:" not in line:
-            continue
-        message = line.split("Parse Error:", 1)[-1].strip()
-        source: str | None = None
-        line_no: int | None = None
-        for look in lines[i : i + 3]:
-            match = _CHECK_SOURCE_RE.search(look)
-            if match:
-                source, line_no = match.group(1), int(match.group(2))
-                break
-        errors.append(ParseError(message=message, source=source, line=line_no))
-    return errors
 
 
 def register_scripts(mcp: FastMCP, bridge: Bridge, config: ServerConfig, runner: Runner) -> None:
